Start splitting backend/src/index.ts into route modules (proof of concept)
index.ts is a 2,438-line monolith with all 49 routes defined inline
inside startServer(). Extracting everything at once is too large to
verify safely in one pass, so this establishes the pattern with a
small, low-risk domain first — matching how the equivalent Tenali
consolidation was sequenced (shared helpers, then one proof-of-concept
route group, then the rest incrementally).

- New backend/src/auth.ts: hoisted getAuthUser, canAccessStudent,
  sanitizeUser, JWT_SECRET/JWT_EXPIRES_IN out of the startServer()
  closure — they were already pure functions of dbStore/req/user, not
  actually dependent on any startServer-local state, so this is a
  behavior-neutral relocation.
- New backend/src/routes/announcements.ts: the 2 announcement routes,
  moved verbatim (registerAnnouncementRoutes(app)), no logic changes.
- index.ts now imports both instead of defining them inline.

Pure reorganization — no route behavior, auth logic, or response
shape changed. Remaining ~47 routes across ~14 domains tracked in a
follow-up issue for incremental extraction.
